Fix bug in set_elt

+++ main.c
@@ -13943,7 +13943,7 @@ set_elt (struct object *seq, unsigned int ind, struct object *val)
 	{
 	  newval = mpz_get_si (val->value_ptr.integer)
 	    & (seq->value_ptr.byte_array->step == 64 ? (unsigned long)-1
-	       : ((1 << seq->value_ptr.byte_array->step)-1));
+	       : (((long) 1 << seq->value_ptr.byte_array->step)-1));
 
 	  switch (seq->value_ptr.byte_array->step)
 	    {
